Changes to pgfree, kthread_end, and LTO support: |- Fixed pgfree assuming direct page mapping when adding page to freelist |- Replaced most assembly in kthread_end into C code |  |- Weird mix that might break with the wrong compilation flags |  |- As a result, implementation will probably be replaced by a reaper thread |- Fixed undefined reference to memset when compiling with LTO

+++ src/proc/lock.h
@@ -4,14 +4,14 @@
 #include <stdint.h>
 
 typedef struct {
-    u8 locked;
+    u32 locked;
 } spinlock_t;
 
 void spinlock(spinlock_t*);
 void spinrelease(spinlock_t*);
 
 typedef struct {
-    u8 locked;
+    u32 locked;
 } sleeplock_t;
 
 void sleeplock(sleeplock_t*);
